Refactor extract phasors
